app/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
fastapi==0.115.11
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.41
langchain-openai==0.3.7
langchain-qdrant==0.2.0
langsmith==0.3.11
numpy==2.2.3
openai==1.65.3
orjson==3.10.15
packaging==24.2
pluggy==1.5.0
portalocker==2.10.1
protobuf==5.29.3
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pytest==8.3.5
python-dotenv==1.0.1
PyYAML==6.0.2
qdrant-client==1.13.3
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.8.2
sniffio==1.3.1
starlette==0.46.0
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
zstandard==0.23.0
